+++ apps/teams-test-app/src/App.tsx
@@ -38,6 +38,7 @@ import PagesCurrentAppAPIs from './components/PagesCurrentAppAPIs';
 import PagesTabsAPIs from './components/PagesTabsAPIs';
 import PeopleAPIs from './components/PeopleAPIs';
 import ChatAPIs from './components/privateApis/ChatAPIs';
+import DataLayerAPIs from './components/privateApis/DataLayerAPIs';
 import ExternalAppAuthenticationAPIs from './components/privateApis/ExternalAppAuthenticationAPIs';
 import ExternalAppCardActionsAPIs from './components/privateApis/ExternalAppCardActionsAPIs';
 import ExternalAppCommandsAPIs from './components/privateApis/ExternalAppCommandsAPIs';
@@ -141,6 +142,7 @@ const App = (): ReactElement => {
         <ChatAPIs />
         <ClipboardAPIs />
         <CustomAPIs />
+        <DataLayerAPIs />
         <DialogAPIs />
         <DialogCardAPIs />
         <DialogCardBotAPIs />
