+++ change/@microsoft-teams-js-f6daee5c-d395-44ed-a4f3-670c3bbc44fe.json
@@ -1,6 +1,6 @@
 {
   "type": "minor",
-  "comment": "Added DataLayer capability that will allow Apps to leverage data layer in the Host. The capability is still awaiting support in one or most host applications. To track availability of this capability across different hosts see https://aka.ms/capmatrix",
+  "comment": "Added private DataLayer capability",
   "packageName": "@microsoft/teams-js",
   "email": "sunatara@microsoft.com",
   "dependentChangeType": "patch"
